add new search bar
