+++ src/components/IllustratedHeaderPageLayout.js
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import useWindowDimensions from '@hooks/useWindowDimensions';
-import useTheme from '../hooks/useTheme';
-import useThemeStyles from '../hooks/useThemeStyles';
+import useTheme from '@hooks/useTheme';
+import useThemeStyles from '@hooks/useThemeStyles';
 import HeaderPageLayout from './HeaderPageLayout';
 import headerWithBackButtonPropTypes from './HeaderWithBackButton/headerWithBackButtonPropTypes';
 import Lottie from './Lottie';
